Activity Log: Extract ActivityCard media previewing into its own component
